Summary:
Although the systemd service was diabled after running for the first time,
it was re-enabled after a package update so every package rebuild would cause
the script to run again.
Now an additional stamp file is created.

	modified:   meta-tplino/recipes-core/firstboot/files/firstboot.sh
	modified:   meta-tplino/recipes-core/firstboot/firstboot.bb

Test Plan:
* read
* build
* deploy VM
* re-build package with bumped PR and update in VM
* reboot VM
* verify that script is not run again

Reviewers: wfailla, aschultz

Reviewed By: aschultz

Subscribers: druss, edogs, mkraemer

Maniphest Tasks: T318

Differential Revision: https://REVIEWSERVER/D1603

+++ meta-tplino/recipes-core/firstboot/files/firstboot.sh
@@ -1,11 +1,17 @@
 #!/bin/sh
 
-if [ -b /dev/sdb ]
+if [ -b /dev/sdb -a ! -b /dev/sdb1 ]
 then
+    # create a new ext4 partition on /dev/sdb using fdisk
     (echo n; echo p; echo ; echo ; echo ; echo w;)|fdisk /dev/sdb
     mkfs.ext4 /dev/sdb1
+    # mount new partition as /srv/data
     mkdir -p /srv/data
     echo "/dev/sdb1  /srv/data  ext4  defaults  0  2" >> /etc/fstab
+    mount -a
+    # disable service
+    systemctl disable firstboot.service
+else
+    echo "Nothing to do."
+    systemctl disable firstboot.service
 fi
-mount -a
-systemctl disable firstboot.service
